requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
defusedxml==0.7.1
Django==4.0
django-allauth==0.47.0
django-currentuser==0.5.3
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==6.0.0
django-simple-history==3.0.0
djangorestframework==3.13.0
idna==3.3
oauthlib==3.1.1
phonenumbers==8.12.39
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
